change listener to visitor
