+++ antlr/pom.xml
@@ -44,6 +44,10 @@
                 <version>${antlr4.version}</version>
                 <executions>
                     <execution>
+                        <configuration>
+                            <listener>false</listener>
+                            <visitor>true</visitor>
+                        </configuration>
                         <goals>
                             <goal>antlr4</goal>
                         </goals>
